Remove OpenSSL from published Alpine images
Previously the Dockerfile's were all installing ca-certificates, curl
and openssl to allow them to securely fetch the glibc apk and the
jre/sdk. However, they were a) not removing openssl after making use of
it and b) including an outdated version (from alpine 3.6 repos
pre-April?) so any downstream images that built upon this official image
are flagged up for various CVEs in 1.0.2n but fixed in 1.0.2o(e.g.,
CVE-2017-3738, CVE-2018-0733, CVE-2018-0739)

```
$ docker run --rm -it ibmjava:8-sfj-alpine apk info -e openssl -d
openssl-1.0.2n-r0 description:
Toolkit for SSL v2/v3 and TLS v1
```

I don't believe there is a need to ship OpenSSL in the image itself, so
this PR switches to adding-and-removing the buildtime dependencies. It
also switches to using LibreSSL as that is what Alpine Linux uses by
default and most OpenSSL CVEs do not apply to it.